add xyz to yaw pitch roll

+++ Tools/Include/Utils.h
@@ -8,7 +8,7 @@ const double pi = CV_PI;
 
 double getPointLength(const cv::Point2f &p);
 
-double get_pitch(float x, float y, float z);
-double get_yaw(float x, float z);
+double get_pitch(double x, double y, double z);
+double get_yaw(double x, double z);
 
 #endif
