+++ Tools/Src/Utils.cpp
@@ -4,11 +4,11 @@ double getPointLength(const cv::Point2f &p) {
     return sqrt(p.x * p.x + p.y * p.y);
 }
 
-double get_pitch(float x, float y, float z) {
+double get_pitch(double x, double y, double z) {
     return atan(y / sqrt(x * x + z * z));
 }
 
-double get_yaw(float x, float z) {
+double get_yaw(double x, double z) {
     if (x == 0 && z == 0)
         return 0;
     else if (x == 0 && z < 0)
